Added a partition key
In line 544, `await this.client.CreateDocumentCollectionIfNotExistsAsync(...)` was called without defining a partition key. This yields errors when you try to add documents later during the demo (line 683/690, where a document is read and created). With the proposed change, the partition key is defined during the creation of the DocumentCollection, so the error won't occurl.